[ADD] modular code with one page website
