--- templates/top.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Your Top Tracks and Artists</title>
-    <style>
-        img {
-            width: 50px;
-            height: 50px;
-            border-radius: 50%;
-            vertical-align: middle;
-            margin-right: 10px;
-        }
-    </style>
-</head>
-<body>
-    <h1>Your Top Tracks and Artists</h1>
-
-    <form method="get" action="/top">
-        <label for="time_range">Choose Time Range:</label>
-        <select name="time_range" id="time_range">
-            <option value="short_term" {% if time_range == 'short_term' %}selected{% endif %}>Last 4 Weeks</option>
-            <option value="medium_term" {% if time_range == 'medium_term' %}selected{% endif %}>Last 6 Months</option>
-            <option value="long_term" {% if time_range == 'long_term' %}selected{% endif %}>All Time</option>
-        </select>
-        <button type="submit">Get Top Tracks & Artists</button>
-    </form>
-
-    <h2>Top Tracks:</h2>
-    <ul>
-        {% for track in top_tracks %}
-        <li>
-            <img src="{{ track.album.images[0].url }}" alt="{{ track.name }} album cover">
-            {{ track.name }} by {{ track.artists[0].name }}
-        </li>
-        {% endfor %}
-    </ul>
-
-    <h2>Top Artists:</h2>
-    <ul>
-        {% for artist in top_artists %}
-            <li>
-                <img src="{{ artist.images[0].url }}" alt="{{ artist.name }}'s image">
-                {{ artist.name }}
-            </li>
-        {% endfor %}
-    </ul>
-</body>
-</html>
